Refactored TopK
