+++ src/main/java/company/vk/polis/ads/TopK.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Comparator;
 import java.util.Iterator;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.stream.Collectors;
 
 /**
@@ -41,69 +40,6 @@ public final class TopK {
         return largest.stream()
                 .sorted(Comparator.reverseOrder())
                 .collect(Collectors.toList());
-    }
-}
-
-class MinHeap<T extends Comparable<T>> {
-    private static final int ROOT_INDEX = 1;
-    private T[] values;
-    private int size;
-
-    public MinHeap(int capacity) {
-        values = (T[]) new Comparable<?>[capacity + 1];
-    }
-
-    public void add(T element) {
-        if (size == values.length - 1) {
-            values = Arrays.copyOf(values, values.length * 3 / 2);
-        }
-        values[++size] = element;
-        swim(size);
-    }
-
-    public T peek() {
-        return values[ROOT_INDEX];
-    }
-
-    public T remove() {
-        T root = values[ROOT_INDEX];
-        swap(ROOT_INDEX, size--);
-        sink(ROOT_INDEX);
-        return root;
-    }
-
-    public int size() {
-        return size;
-    }
-
-    private void swap(int firstIndex, int secondIndex) {
-        T temp = values[firstIndex];
-        values[firstIndex] = values[secondIndex];
-        values[secondIndex] = temp;
-    }
-
-    private void swim(int index) {
-        int currIndex = index;
-        while (currIndex > 1 && values[currIndex].compareTo(values[currIndex / 2]) < 0) {
-            swap(currIndex, currIndex / 2);
-            currIndex /= 2;
-        }
-    }
-
-    private void sink(int index) {
-        int currIndex = index;
-        int childIndex;
-        while (2 * currIndex <= size) {
-            childIndex = 2 * currIndex;
-            if (childIndex < size && values[childIndex].compareTo(values[childIndex + 1]) > 0) {
-                childIndex++;
-            }
-            if (values[currIndex].compareTo(values[childIndex]) <= 0) {
-                break;
-            }
-            swap(childIndex, currIndex);
-            currIndex = childIndex;
-        }
     }
 }
